Lots of changes.
Added Poison
Added Health Bar
Towers can be upgraded.
Simplified upgrading/costs on my end.

+++ enemies/Enemy.as
@@ -5,10 +5,12 @@
 	import flash.display.Sprite;
 	import flash.events.*;
 	import flash.geom.Point;
+	import design.HealthBar;
 
 
 	public class Enemy extends Sprite
 	{
+		internal var eMaxHp:int;
 		internal var eHp:int;
 		internal var pt:Point = new Point();
 		public var maxMoveSpeed:Number;
@@ -23,19 +25,33 @@
 		public var distanceTraveled:Number;
 		public var killed:Boolean;
 
+		public var poisonSlow:Number;
+		public var iceSlow:Number;
+		private var healthBar:HealthBar;
+
 		public function Enemy(Map:Array)
 		{
+			poisonSlow = 0;
+			iceSlow = 0;
 			distanceTraveled = 0;
 			id = 9999 * Math.random();
 			mapArray = Map;
+			healthBar = new HealthBar();
 			addEventListener(Event.ADDED_TO_STAGE, added);
 			addEventListener(Event.ENTER_FRAME, startMovement);
 			// constructor code
 		}
 		private function added(e:Event):void
 		{
+			eMaxHp = eHp;
 			_root = MovieClip(root);
+			_root.addChild(healthBar);
+			healthBar.visible = false;
 			removeEventListener(Event.ADDED_TO_STAGE, added);
+		}
+		public function determineMoveSpeed():void
+		{
+			moveSpeed = maxMoveSpeed - (maxMoveSpeed * (iceSlow + poisonSlow));
 		}
 		internal function determineArmor(dmg:int, damageType:String):Number
 		{
@@ -76,15 +92,33 @@
 			if (eHp <= 0)
 			{
 				killed = true;
-				destroyThis();
+				if (_root != null && _root.contains(this))
+				{
+					destroyThis();
+				}
 			}
 		}
 		internal function startMovement(e:Event):void
 		{
 			moveCharacter();
 		}
+		public function healthBarOnOff():void
+		{
+			healthBar.visible = !(healthBar.visible);
+		}
+		private function updateHealth():void
+		{
+			var percentHP:Number = eHp / eMaxHp;
+			if (_root != null && _root.contains(healthBar))
+			{
+				healthBar.hpBar.scaleX = percentHP;
+				healthBar.x = this.x;
+				healthBar.y = this.y;
+			}
+		}
 		internal function moveCharacter():void
 		{
+
 			switch (mapArray[Math.floor(pt.y)][Math.floor(pt.x)])
 			{
 				case 1 :
@@ -94,7 +128,7 @@
 					{
 						distanceTraveled -=  x % 32;
 						x -=  x % 32;
-						
+
 						pt.x = x / 32;
 					}
 					break;
@@ -106,7 +140,7 @@
 					{
 						distanceTraveled -=  y % 32;
 						y -=  y % 32;
-						
+
 						pt.y = y / 32;
 
 					}
@@ -117,11 +151,11 @@
 					x -=  moveSpeed;
 					if (Math.ceil(x/32) != pt.x)
 					{
-						
+
 						pt.x--;
 						distanceTraveled -=  Math.abs(pt.x * 32 - x);
 						x +=  Math.abs(pt.x * 32 - x);
-						
+
 					}
 					break;
 
@@ -130,11 +164,11 @@
 
 					if (Math.ceil(y/32) != pt.y)
 					{
-						
+
 						pt.y--;
 						distanceTraveled -=  Math.abs(pt.y * 32 - y);
 						y +=  Math.abs(pt.y * 32 - y);
-						
+
 					}
 					break;
 
@@ -143,12 +177,15 @@
 					break;
 			}
 			distanceTraveled +=  moveSpeed;
+			updateHealth();
 		}
 
 		private function destroyThis():void
 		{
 			mapArray = null;
 			removeEventListener(Event.ENTER_FRAME, startMovement);
+			_root.removeChild(healthBar);
+			healthBar = null;
 			_root.removeChild(this);
 			_root = null;
 		}
